Proper kill

+++ scripts/gstart.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-pkill -f "kitty -o allow_remote_control=yes --listen-on unix:/tmp/kitty_sock -T local" || :
+if [[ "$(hostname)" == "abx1gen3" ]]; then
+    killwork
+else
+    pkill -f "kitty -o allow_remote_control=yes --listen-on unix:/tmp/kitty_sock -T local" || :
+fi
 
 kitty -o allow_remote_control=yes --listen-on unix:/tmp/kitty_sock -T local "$HOME/scripts/tstart.sh" local &
